new components
git-svn-id: http://svn.code.sf.net/p/lazarus-ccr/svn@704 8e941d3f-bd1b-0410-a28a-d453659cc2b4

+++ components/rx/rxtimeedit.pas
@@ -140,7 +140,7 @@ end;
 procedure TCustomRxTimeEdit.DoChangeValue(AValue: integer);
 var
   S:ShortString;
-  H1, M2, S3:byte;
+  H1, M2, S3:Integer;
   i,p:integer;
 
 procedure IncHour;
